feat(kuma-cp) rename Local to Remote
